Re-implement grid rendering + added more useful ToolHelper methods
- Start amending existing tool damage/checking implementations
- BlockItemPipe now harvestable with wrench instead of pickaxe
