+++ src/main/java/gregtech/common/items/ToolItems.java
@@ -1,12 +1,22 @@
 package gregtech.common.items;
 
+import codechicken.lib.vec.Vector3;
 import gregtech.api.GTValues;
+import gregtech.api.capability.GregtechTileCapabilities;
+import gregtech.api.cover.CoverDefinition;
+import gregtech.api.cover.ICoverable;
 import gregtech.api.items.toolitem.IGTTool;
 import gregtech.api.items.toolitem.ItemGTTool;
 import gregtech.api.items.toolitem.ToolBuilder;
 import gregtech.api.items.toolitem.ToolHelper;
+import gregtech.api.metatileentity.MetaTileEntity;
+import gregtech.api.metatileentity.MetaTileEntityHolder;
+import gregtech.api.pipenet.block.BlockPipe;
+import gregtech.api.pipenet.tile.IPipeTile;
+import gregtech.api.pipenet.tile.TileEntityPipeBase;
 import gregtech.api.sound.GTSounds;
 import gregtech.api.unification.OreDictUnifier;
+import gregtech.api.util.GTUtility;
 import gregtech.api.util.TaskScheduler;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.Minecraft;
@@ -21,6 +31,8 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.util.math.Vec3d;
@@ -44,6 +56,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.ListIterator;
 import java.util.Set;
+import java.util.function.BooleanSupplier;
+import java.util.function.Predicate;
 
 public class ToolItems {
 
@@ -264,18 +278,27 @@ public class ToolItems {
     }
 
     // Handle client-view of harvestable blocks in AoE (and potentially wrench overlay in the future)
+    // Handle machine grid rendering as well
     @SubscribeEvent
     @SideOnly(Side.CLIENT)
     public static void onDrawHighlightEvent(DrawBlockHighlightEvent event) {
         EntityPlayer player = event.getPlayer();
-        if (!player.isSneaking()) {
-            ItemStack stack = player.getHeldItemMainhand();
-            Item item = stack.getItem();
-            if (item instanceof IGTTool) {
-                Set<BlockPos> validPositions = ((IGTTool) item).getHarvestableBlocks(stack, player.world, player, event.getTarget());
+        ItemStack stack = player.getHeldItemMainhand();
+        BlockPos pos = event.getTarget().getBlockPos();
+        IBlockState state = player.world.getBlockState(pos);
+        TileEntity tile = player.world.getTileEntity(event.getTarget().getBlockPos());
+        boolean sneaking = player.isSneaking();
+        if (shouldRenderGridOverlays(state, tile, stack, player.getHeldItemOffhand(), sneaking) &&
+                renderGridOverlays(player, pos, state, event.getTarget().sideHit, tile, event.getPartialTicks())) {
+            event.setCanceled(true);
+            return;
+        }
+        if (!sneaking) {
+            if (stack.getItem() instanceof IGTTool) {
+                Set<BlockPos> validPositions = ((IGTTool) stack.getItem()).getHarvestableBlocks(stack, player.world, player, event.getTarget());
                 float partialTicks = event.getPartialTicks();
-                for (BlockPos pos : validPositions) {
-                    event.getContext().drawSelectionBox(player, new RayTraceResult(Vec3d.ZERO, null, pos), 0, partialTicks);
+                for (BlockPos validPosition : validPositions) {
+                    event.getContext().drawSelectionBox(player, new RayTraceResult(Vec3d.ZERO, null, validPosition), 0, partialTicks);
                 }
                 DestroyBlockProgress progress = event.getContext().damagedBlocks.get(player.getEntityId());
                 if (progress != null) {
@@ -291,12 +314,12 @@ public class ToolItems {
                         bufferBuilder.begin(GL11.GL_QUADS, DefaultVertexFormats.BLOCK);
                         bufferBuilder.setTranslation(-relX, -relY, -relZ);
                         bufferBuilder.noColor(); // ?
-                        for (BlockPos pos : validPositions) {
-                            TileEntity tileEntity = mc.world.getTileEntity(pos);
+                        for (BlockPos validPosition : validPositions) {
+                            TileEntity tileEntity = mc.world.getTileEntity(validPosition);
                             boolean hasBreak = tileEntity != null && tileEntity.canRenderBreaking();
                             if (!hasBreak) {
                                 TextureAtlasSprite sprite = event.getContext().destroyBlockIcons[damage];
-                                rendererDispatcher.renderBlockDamage(mc.world.getBlockState(pos), pos, sprite, mc.world);
+                                rendererDispatcher.renderBlockDamage(mc.world.getBlockState(validPosition), validPosition, sprite, mc.world);
                             }
                         }
                         Tessellator.getInstance().draw();
@@ -326,6 +349,302 @@ public class ToolItems {
         GlStateManager.enableAlpha();
         GlStateManager.depthMask(true);
         GlStateManager.popMatrix();
+    }
+
+    private static boolean shouldRenderGridOverlays(IBlockState state, TileEntity tile, ItemStack mainHand, ItemStack offHand, boolean isSneaking) {
+        if (state.getBlock() instanceof BlockPipe) {
+            BlockPipe<?, ?, ?> pipe = (BlockPipe<?, ?, ?>) state.getBlock();
+            if (isSneaking && mainHand.getItem().getClass() == Item.getItemFromBlock(pipe).getClass()) {
+                return true;
+            } else {
+                if (mainHand.getItem().getToolClasses(mainHand).stream().anyMatch(s -> s.equals("screwdriver") || pipe.isToolEffective(s, state))) {
+                    return true;
+                } else if (offHand.getItem().getToolClasses(offHand).stream().anyMatch(s -> s.equals("screwdriver") || pipe.isToolEffective(s, state))) {
+                    return true;
+                }
+                BooleanSupplier hasCover = () -> tile instanceof IPipeTile && ((IPipeTile<?, ?>) tile).getCoverableImplementation().hasAnyCover();
+                Predicate<CoverDefinition> canCover = coverDef -> tile instanceof IPipeTile && ICoverable.canPlaceCover(coverDef, ((IPipeTile<?, ?>) tile).getCoverableImplementation());
+                if (GTUtility.isCoverBehaviorItem(mainHand, hasCover, canCover) || GTUtility.isCoverBehaviorItem(offHand, hasCover, canCover)) {
+                    return true;
+                }
+            }
+        }
+        if (tile instanceof MetaTileEntityHolder) {
+            MetaTileEntity mte = ((MetaTileEntityHolder) tile).getMetaTileEntity();
+            if (mte != null && mte.canRenderMachineGrid() &&
+                    (mainHand.getItem().getToolClasses(mainHand).contains("wrench") ||
+                            offHand.getItem().getToolClasses(offHand).contains("wrench"))) {
+                return true;
+            }
+        }
+        ICoverable coverable = tile.getCapability(GregtechTileCapabilities.CAPABILITY_COVERABLE, null);
+        return coverable != null && GTUtility.isCoverBehaviorItem(mainHand, coverable::hasAnyCover, coverDef -> ICoverable.canPlaceCover(coverDef, coverable));
+    }
+
+    private static float rColour;
+    private static float gColour;
+    private static float bColour;
+
+    private static boolean renderGridOverlays(EntityPlayer player, BlockPos pos, IBlockState state, EnumFacing facing, TileEntity tile, float partialTicks) {
+        if (player.world.getWorldBorder().contains(pos)) {
+            GlStateManager.enableBlend();
+            GlStateManager.tryBlendFuncSeparate(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
+            GlStateManager.glLineWidth(2.0F);
+            GlStateManager.disableTexture2D();
+            GlStateManager.depthMask(false);
+            double d3 = player.lastTickPosX + (player.posX - player.lastTickPosX) * (double) partialTicks;
+            double d4 = player.lastTickPosY + (player.posY - player.lastTickPosY) * (double) partialTicks;
+            double d5 = player.lastTickPosZ + (player.posZ - player.lastTickPosZ) * (double) partialTicks;
+            AxisAlignedBB box = state.getSelectedBoundingBox(player.world, pos).grow(0.002D).offset(-d3, -d4, -d5);
+            RenderGlobal.drawSelectionBoundingBox(box, 1, 1, 1, 0.4F);
+
+            rColour = gColour = bColour = 0.2F + (float) Math.sin((float) (System.currentTimeMillis() % (Math.PI * 800)) / 800) / 2;
+
+            if (tile instanceof TileEntityPipeBase) {
+                TileEntityPipeBase<?, ?> tepb = (TileEntityPipeBase<?, ?>) tile;
+                drawGridOverlays(facing, box, face -> tepb.isConnected(face) || tepb.getCoverableImplementation().getCoverAtSide(face) != null);
+            } else if (tile instanceof MetaTileEntityHolder) {
+                MetaTileEntity mte = ((MetaTileEntityHolder) tile).getMetaTileEntity();
+                drawGridOverlays(facing, box, mte::isSideUsed);
+            } else {
+                drawGridOverlays(box);
+            }
+            GlStateManager.depthMask(true);
+            GlStateManager.enableTexture2D();
+            GlStateManager.disableBlend();
+            return true;
+        }
+        return false;
+    }
+
+    private static void drawGridOverlays(AxisAlignedBB box) {
+        Tessellator tessellator = Tessellator.getInstance();
+        BufferBuilder buffer = tessellator.getBuffer();
+        buffer.begin(3, DefaultVertexFormats.POSITION_COLOR);
+
+        Vector3 topRight = new Vector3(box.maxX, box.maxY, box.maxZ);
+        Vector3 bottomRight = new Vector3(box.maxX, box.minY, box.maxZ);
+        Vector3 bottomLeft = new Vector3(box.minX, box.minY, box.maxZ);
+        Vector3 topLeft = new Vector3(box.minX, box.maxY, box.maxZ);
+        Vector3 shift = new Vector3(0.25, 0, 0);
+        Vector3 shiftVert = new Vector3(0, 0.25, 0);
+
+        Vector3 cubeCenter = new Vector3(box.getCenter());
+
+        topRight.subtract(cubeCenter);
+        bottomRight.subtract(cubeCenter);
+        bottomLeft.subtract(cubeCenter);
+        topLeft.subtract(cubeCenter);
+
+        topRight.add(cubeCenter);
+        bottomRight.add(cubeCenter);
+        bottomLeft.add(cubeCenter);
+        topLeft.add(cubeCenter);
+
+        // straight top bottom lines
+        startLine(buffer, topRight.copy().add(shift.copy().negate()));
+        endLine(buffer, bottomRight.copy().add(shift.copy().negate()));
+
+        startLine(buffer, bottomLeft.copy().add(shift));
+        endLine(buffer, topLeft.copy().add(shift));
+
+        // straight side to side lines
+        startLine(buffer, topLeft.copy().add(shiftVert.copy().negate()));
+        endLine(buffer, topRight.copy().add(shiftVert.copy().negate()));
+
+        startLine(buffer, bottomLeft.copy().add(shiftVert));
+        endLine(buffer, bottomRight.copy().add(shiftVert));
+
+        tessellator.draw();
+    }
+
+    private static void drawGridOverlays(EnumFacing facing, AxisAlignedBB box, Predicate<EnumFacing> test) {
+        Tessellator tessellator = Tessellator.getInstance();
+        BufferBuilder buffer = tessellator.getBuffer();
+        buffer.begin(3, DefaultVertexFormats.POSITION_COLOR);
+
+        Vector3 topRight = new Vector3(box.maxX, box.maxY, box.maxZ);
+        Vector3 bottomRight = new Vector3(box.maxX, box.minY, box.maxZ);
+        Vector3 bottomLeft = new Vector3(box.minX, box.minY, box.maxZ);
+        Vector3 topLeft = new Vector3(box.minX, box.maxY, box.maxZ);
+        Vector3 shift = new Vector3(0.25, 0, 0);
+        Vector3 shiftVert = new Vector3(0, 0.25, 0);
+
+        Vector3 cubeCenter = new Vector3(box.getCenter());
+
+        topRight.subtract(cubeCenter);
+        bottomRight.subtract(cubeCenter);
+        bottomLeft.subtract(cubeCenter);
+        topLeft.subtract(cubeCenter);
+
+        boolean leftBlocked;
+        boolean topBlocked;
+        boolean rightBlocked;
+        boolean bottomBlocked;
+        boolean frontBlocked = test.test(facing);
+        boolean backBlocked = test.test(facing.getOpposite());
+
+        switch (facing) {
+            case WEST: {
+                topRight.rotate(Math.PI / 2, Vector3.down);
+                bottomRight.rotate(Math.PI / 2, Vector3.down);
+                bottomLeft.rotate(Math.PI / 2, Vector3.down);
+                topLeft.rotate(Math.PI / 2, Vector3.down);
+                shift.rotate(Math.PI / 2, Vector3.down);
+                shiftVert.rotate(Math.PI / 2, Vector3.down);
+
+                leftBlocked = test.test(EnumFacing.NORTH);
+                topBlocked = test.test(EnumFacing.UP);
+                rightBlocked = test.test(EnumFacing.SOUTH);
+                bottomBlocked = test.test(EnumFacing.DOWN);
+                break;
+            }
+            case EAST: {
+                topRight.rotate(-Math.PI / 2, Vector3.down);
+                bottomRight.rotate(-Math.PI / 2, Vector3.down);
+                bottomLeft.rotate(-Math.PI / 2, Vector3.down);
+                topLeft.rotate(-Math.PI / 2, Vector3.down);
+                shift.rotate(-Math.PI / 2, Vector3.down);
+                shiftVert.rotate(-Math.PI / 2, Vector3.down);
+
+                leftBlocked = test.test(EnumFacing.SOUTH);
+                topBlocked = test.test(EnumFacing.UP);
+                rightBlocked = test.test(EnumFacing.NORTH);
+                bottomBlocked = test.test(EnumFacing.DOWN);
+                break;
+            }
+            case NORTH: {
+                topRight.rotate(Math.PI, Vector3.down);
+                bottomRight.rotate(Math.PI, Vector3.down);
+                bottomLeft.rotate(Math.PI, Vector3.down);
+                topLeft.rotate(Math.PI, Vector3.down);
+                shift.rotate(Math.PI, Vector3.down);
+                shiftVert.rotate(Math.PI, Vector3.down);
+
+                leftBlocked = test.test(EnumFacing.EAST);
+                topBlocked = test.test(EnumFacing.UP);
+                rightBlocked = test.test(EnumFacing.WEST);
+                bottomBlocked = test.test(EnumFacing.DOWN);
+                break;
+            }
+            case UP: {
+                Vector3 side = new Vector3(1, 0, 0);
+                topRight.rotate(-Math.PI / 2, side);
+                bottomRight.rotate(-Math.PI / 2, side);
+                bottomLeft.rotate(-Math.PI / 2, side);
+                topLeft.rotate(-Math.PI / 2, side);
+                shift.rotate(-Math.PI / 2, side);
+                shiftVert.rotate(-Math.PI / 2, side);
+
+                leftBlocked = test.test(EnumFacing.WEST);
+                topBlocked = test.test(EnumFacing.NORTH);
+                rightBlocked = test.test(EnumFacing.EAST);
+                bottomBlocked = test.test(EnumFacing.SOUTH);
+                break;
+            }
+            case DOWN: {
+                Vector3 side = new Vector3(1, 0, 0);
+                topRight.rotate(Math.PI / 2, side);
+                bottomRight.rotate(Math.PI / 2, side);
+                bottomLeft.rotate(Math.PI / 2, side);
+                topLeft.rotate(Math.PI / 2, side);
+                shift.rotate(Math.PI / 2, side);
+                shiftVert.rotate(Math.PI / 2, side);
+
+                leftBlocked = test.test(EnumFacing.WEST);
+                topBlocked = test.test(EnumFacing.SOUTH);
+                rightBlocked = test.test(EnumFacing.EAST);
+                bottomBlocked = test.test(EnumFacing.NORTH);
+                break;
+            }
+            default: {
+                leftBlocked = test.test(EnumFacing.WEST);
+                topBlocked = test.test(EnumFacing.UP);
+                rightBlocked = test.test(EnumFacing.EAST);
+                bottomBlocked = test.test(EnumFacing.DOWN);
+            }
+        }
+
+        topRight.add(cubeCenter);
+        bottomRight.add(cubeCenter);
+        bottomLeft.add(cubeCenter);
+        topLeft.add(cubeCenter);
+
+        // straight top bottom lines
+        startLine(buffer, topRight.copy().add(shift.copy().negate()));
+        endLine(buffer, bottomRight.copy().add(shift.copy().negate()));
+
+        startLine(buffer, bottomLeft.copy().add(shift));
+        endLine(buffer, topLeft.copy().add(shift));
+
+        // straight side to side lines
+        startLine(buffer, topLeft.copy().add(shiftVert.copy().negate()));
+        endLine(buffer, topRight.copy().add(shiftVert.copy().negate()));
+
+        startLine(buffer, bottomLeft.copy().add(shiftVert));
+        endLine(buffer, bottomRight.copy().add(shiftVert));
+
+        if (leftBlocked) {
+            startLine(buffer, topLeft.copy().add(shiftVert.copy().negate()));
+            endLine(buffer, bottomLeft.copy().add(shiftVert.copy()).add(shift));
+
+            startLine(buffer, topLeft.copy().add(shiftVert.copy().negate()).add(shift));
+            endLine(buffer, bottomLeft.copy().add(shiftVert));
+        }
+        if (topBlocked) {
+            startLine(buffer, topLeft.copy().add(shift));
+            endLine(buffer, topRight.copy().add(shift.copy().negate()).add(shiftVert.copy().negate()));
+
+            startLine(buffer, topLeft.copy().add(shift).add(shiftVert.copy().negate()));
+            endLine(buffer, topRight.copy().add(shift.copy().negate()));
+        }
+        if (rightBlocked) {
+            startLine(buffer, topRight.copy().add(shiftVert.copy().negate()));
+            endLine(buffer, bottomRight.copy().add(shiftVert.copy()).add(shift.copy().negate()));
+
+            startLine(buffer, topRight.copy().add(shiftVert.copy().negate()).add(shift.copy().negate()));
+            endLine(buffer, bottomRight.copy().add(shiftVert));
+        }
+        if (bottomBlocked) {
+            startLine(buffer, bottomLeft.copy().add(shift));
+            endLine(buffer, bottomRight.copy().add(shift.copy().negate()).add(shiftVert));
+
+            startLine(buffer, bottomLeft.copy().add(shift).add(shiftVert));
+            endLine(buffer, bottomRight.copy().add(shift.copy().negate()));
+        }
+        if (frontBlocked) {
+            startLine(buffer, topLeft.copy().add(shift).add(shiftVert.copy().negate()));
+            endLine(buffer, bottomRight.copy().add(shift.copy().negate()).add(shiftVert));
+
+            startLine(buffer, topRight.copy().add(shift.copy().negate()).add(shiftVert.copy().negate()));
+            endLine(buffer, bottomLeft.copy().add(shift).add(shiftVert));
+        }
+        if (backBlocked) {
+            Vector3 localXShift = new Vector3(0, 0, 0); // Set up translations for the current X.
+            for (int i = 0; i < 2; i++) {
+                Vector3 localXShiftVert = new Vector3(0, 0, 0);
+                for (int j = 0; j < 2; j++) {
+                    startLine(buffer, topLeft.copy().add(localXShift).add(localXShiftVert));
+                    endLine(buffer, topLeft.copy().add(localXShift).add(localXShiftVert).add(shift).subtract(shiftVert));
+
+                    startLine(buffer, topLeft.copy().add(localXShift).add(localXShiftVert).add(shift));
+                    endLine(buffer, topLeft.copy().add(localXShift).add(localXShiftVert).subtract(shiftVert));
+
+                    localXShiftVert.add(bottomLeft.copy().subtract(topLeft).add(shiftVert)); // Move by the vector from the top to the bottom, minus the shift from the edge.
+                }
+                localXShift.add(topRight.copy().subtract(topLeft).subtract(shift)); // Move by the vector from the left to the right, minus the shift from the edge.
+            }
+        }
+
+        tessellator.draw();
+    }
+
+    private static void startLine(BufferBuilder buffer, Vector3 vec) {
+        buffer.pos(vec.x, vec.y, vec.z).color(rColour, gColour, bColour, 0.0F).endVertex();
+    }
+
+    private static void endLine(BufferBuilder buffer, Vector3 vec) {
+        buffer.pos(vec.x, vec.y, vec.z).color(rColour, gColour, bColour, 1F).endVertex();
     }
 
 }
