--- src/main/java/gregtech/common/tools/DamageValues.java
@@ -1,13 +1,0 @@
-package gregtech.common.tools;
-
-public class DamageValues {
-    public static final int DAMAGE_FOR_SCREWDRIVER = 1;
-    public static final int DAMAGE_FOR_WRENCH = 2;
-    public static final int DAMAGE_FOR_CUTTER = 2;
-    public static final int DAMAGE_FOR_CROWBAR = 1;
-    public static final int DAMAGE_FOR_SOFT_HAMMER = 3;
-    public static final int DAMAGE_FOR_HAMMER = 3;
-    public static final int DAMAGE_FOR_HOE = 2;
-    public static final int DAMAGE_FOR_PLUNGER = 1;
-    public static final int DAMAGE_FOR_SHOVEL = 1;
-}
